moved title fetch for h5p to h5pApi.js
